Validations fr inventory managament
May binago ko kung paano mag process ang pgacreate ng bouquet sa server side

+++ app/Http/Controllers/OrderManagementController.php
@@ -179,11 +179,11 @@ class OrderManagementController extends Controller
 
     public function DeleteAcessory_per_SessionBqt_Order($Acessory_ID)
 	{
-        /*if(auth::check() == false){
-            Session::put('loginSession','fail');
-            return redirect() -> route('adminsignin');
-        }
-        else{*/
+    if(auth::check() == false){
+        Session::put('loginSession','fail');
+        return redirect() -> route('adminsignin');
+    }
+    else{
 			echo $Acessory_ID;
 			foreach(Cart::instance('OrderedBqt_Acessories')->content() as $row){
 				if($row->id == $Acessory_ID){
@@ -192,10 +192,7 @@ class OrderManagementController extends Controller
 				}
 			}
       return redirect()-> route('Long_Sales_Order.index');
-
-	     //   return redirect()->route('Order.CustomizeaBouquet');
-			//returns to the creation of bouquet*/
-    	//}
+    	}
 	}//end of function
 
 
@@ -236,11 +233,11 @@ class OrderManagementController extends Controller
 
 	public function Cancel_and_Clear_BqtSession_Order()
 	{
-        /*if(auth::check() == false){
-            Session::put('loginSession','fail');
-            return redirect() -> route('adminsignin');
-        }
-        else{*/
+    if(auth::check() == false){
+        Session::put('loginSession','fail');
+        return redirect() -> route('adminsignin');
+    }
+    else{
 	        $AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
 
 			Cart::instance('OrderedBqt_Flowers')->destroy();
@@ -251,18 +248,17 @@ class OrderManagementController extends Controller
 			     return view('Orders.creationOfOrders')
 	             ->with('FlowerList',$AvailableFlowers);
 			return redirect()->route('Sales_Qoutation.show');//returns to the creation of flowers
-		//}
+		}
 	}//end of function
 
 
     public function Clear_Cart()
     {
-          /*if(auth::check() == false){
-              Session::put('loginSession','fail');
-              return redirect() -> route('adminsignin');
-          }
-          else{*/
-
+     if(auth::check() == false){
+          Session::put('loginSession','fail');
+          return redirect() -> route('adminsignin');
+      }
+      else{
         Cart::instance('FinalBqt_Flowers')->destroy();
         Cart::instance('FinalBqt_Acessories')->destroy();
         Cart::instance('Ordered_Bqt')->destroy();
@@ -271,17 +267,17 @@ class OrderManagementController extends Controller
         Session::put('CartClearSession', 'Successful');
                 return redirect()-> route('Long_Sales_Order.index');
         //returns to the creation of flowers
-      //}
+      }
     }//end of function
 
 
     public function Clear_Bouquet()
     {
-          /*if(auth::check() == false){
-              Session::put('loginSession','fail');
-              return redirect() -> route('adminsignin');
-          }
-          else{*/
+      if(auth::check() == false){
+          Session::put('loginSession','fail');
+          return redirect() -> route('adminsignin');
+      }
+      else{
 
         Cart::instance('OrderedBqt_Flowers')->destroy();
         Cart::instance('OrderedBqt_Acessories')->destroy();
@@ -289,16 +285,16 @@ class OrderManagementController extends Controller
         Session::put('BqtClearSession', 'Successful');
                 return redirect()-> route('Long_Sales_Order.index');
         //returns to the creation of flowers
-      //}
+      }
     }//end of function
 
 	public function saveCustomized_Bqt($order_ID)
 	{
-        /*if(auth::check() == false){
-            Session::put('loginSession','fail');
-            return redirect() -> route('adminsignin');
-        }
-        else{*/
+    if(auth::check() == false){
+          Session::put('loginSession','fail');
+          return redirect() -> route('adminsignin');
+      }
+      else{
 				$BQT_Flower_Count = 0;
 				foreach(Cart::instance('OrderedBqt_Flowers')->content() as $row){
 						$BQT_Flower_Count += $row->qty;
@@ -352,23 +348,19 @@ class OrderManagementController extends Controller
 
 			Session::put('Save_Bouqet_To_myOrder', 'Successful');
 			return redirect()->route('Sales_Qoutation.show', $order_ID);//returns to the creation of flowers
-		//}
+		}
 	}//end of function
 
 
 
 
 	public function saveNewCustomized_Bqt(){
-        /*if(auth::check() == false){
+        if(auth::check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 			$AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
-
-			/*Cart::instance('FinalBqt_Acessories')->destroy();
-			Cart::instance('Ordered_Bqt')->destroy();
-			Cart::instance('FinalBqt_Flowers')->destroy();*/
 
 				$BQT_Flower_Count = 0;
 				foreach(Cart::instance('OrderedBqt_Flowers')->content() as $row){
@@ -378,11 +370,6 @@ class OrderManagementController extends Controller
 			    $Flowers_subtotal = str_replace(array(','), array(''), Cart::instance('OrderedBqt_Flowers')->subtotal());
 
 	            $Acessories_subtotal = str_replace(array(','), array(''), Cart::instance('OrderedBqt_Acessories')->subtotal());
-	 /*           echo '<h2><b>$Flowers_subtotal = </b>'.$Flowers_subtotal.'<\h2>';
-	            echo '<h2><b>$Acessories_subtotal = </b>'.$Acessories_subtotal.'<\h2>';
-	            echo '<h2><b>BOQUET_total = </b>'.number_format($Acessories_subtotal + $Flowers_subtotal,2).'<\h2>';
-	            echo '<h2><b>BOQUET Flower Count = </b>'.$BQT_Flower_Count.' pcs. <\h2>';
-	*/
 
 	            $BQT_Price = $Acessories_subtotal + $Flowers_subtotal;
 
@@ -392,17 +379,6 @@ class OrderManagementController extends Controller
 						Cart::instance('Ordered_Bqt')
 				         ->add(['id' => $bqt_Id, 'name' => $bqtName, 'qty' => 1, 'price' => $BQT_Price,
 				         		'options' => ['count' => $BQT_Flower_Count]]);
-						/*
-				    	foreach(Cart::instance('Ordered_Bqt')->content() as $bqt){
-				    		echo '<div class = "row">';
-				    		echo '<h1>The Bouquet Detais<h1><hr><br>';
-				        	echo '<h2><b> $bqt_ID = </b>'.$bqt->id.'</h2>';
-				        	echo '<h4><b> $bqt_name = </b>'.$bqt->name.'</h4>';
-				        	echo '<h4><b> $bqt_qty = </b>'.$bqt->qty.'</h4>';
-				        	echo '<h4><b> $bqt_price = </b>'.$bqt->price.'</h4>';
-				        	echo '<h4><b> $bqt_count = </b>'.$bqt->options['count'].'</h4>';
-				        	echo '</div>';
-				    	}  */
 
 				        foreach(Cart::instance('OrderedBqt_Flowers')->content() as $row){
 						    //this foreach will transafer all of their content to another session
@@ -411,51 +387,10 @@ class OrderManagementController extends Controller
                       'priceType'=>$row->options['priceType'], 'bqt_ID' => $bqt_Id]]);
 				        }//END OF INNER FOREACH of flower cart
 
-				       /* foreach(Cart::instance('FinalBqt_Flowers')->content() as $row){
-				        	echo '<div class = "row">';
-				        	echo '<div class = "col-md-5">';
-				        	echo '<h2><b> Flwr_ID = </b>'.$row->id.'</h2>';
-				        	echo '<h4><b> Flwr_name = </b>'.$row->name.'</h4>';
-				        	echo '<h4><b> Flwr_qty = </b>'.$row->qty.'</h4>';
-				        	echo '<h4><b> Flwr_price = </b>'.$row->price.'</h4>';
-				        	echo '</div>';
-				        	echo '<div class = "col-md-6">';
-				        	echo '<h2><b> Flwr_o_price = </b>'.$row->options['orig_price'].'</h2>';
-				        	echo '<h4><b> Flwr_image = </b>'.$row->options['image'].'</h4>';
-				        	echo '<h4><b> Flwr_type = </b>'.$row->options['priceType'].'</h4>';
-				        	echo '<h4><b> Flwr_T_Amt = </b>'.$row->options['T_Amt'].'</h4>';
-				        	echo '<h4><b> Flwr_bqtID = </b>'.$row->options['bqt_ID'].'</h4>';
-				        	echo '</div>';
-				        	echo '</div>';
-				        	echo '<hr><br>';
-				        }
-				        echo '<br><hr><br>';*/
 				        foreach(Cart::instance('OrderedBqt_Acessories')->content() as $Acrow){
 					        Cart::instance('FinalBqt_Acessories')
 		                		->add(['id' => $Acrow->id, 'name' => $Acrow->name, 'qty' => $Acrow->qty, 'price' => $Acrow->price,'options' => ['orig_price' => $Acrow->options['orig_price'],'T_Amt' => $Acrow->options['T_Amt'],'image'=>$Acrow->options['image'],'priceType'=>$Acrow->options['priceType'],'bqt_ID' => $bqt_Id]]);
 				        }//end of foreach of the acessories cart
-
-
-				        /*
-				        foreach(Cart::instance('FinalBqt_Acessories')->content() as $Acrow2){
-				        	echo '<div class = "row">';
-				        	echo '<div class = "col-md-5">';
-				        	echo '<h2><b> Acrs_ID = </b>'.$Acrow2->id.'</h2>';
-				        	echo '<h4><b> Acrs_name = </b>'.$Acrow2->name.'</h4>';
-				        	echo '<h4><b> Acrs_qty = </b>'.$Acrow2->qty.'</h4>';
-				        	echo '<h4><b> Acrs_price = </b>'.$Acrow2->price.'</h4>';
-				        	echo '</div>';
-				        	echo '<div class = "col-md-6">';
-				        	echo '<h2><b> Acrs_o_price = </b>'.$Acrow2->options['orig_price'].'</h2>';
-				        	echo '<h4><b> Acrs_image = </b>'.$Acrow2->options['image'].'</h4>';
-				        	echo '<h4><b> Acrs_type = </b>'.$Acrow2->options['priceType'].'</h4>';
-				        	echo '<h4><b> Acrs_T_Amt = </b>'.$Acrow2->options['T_Amt'].'</h4>';
-				        	echo '<h4><b> Acrs_bqtID = </b>'.$Acrow2->options['bqt_ID'].'</h4>';
-				        	echo '</div>';
-				        	echo '</div>';
-				        	echo '<hr><br>';
-				        }
-				    	*/
 				}//end of if
 				else{
 					$newBqt_Id = '';
@@ -467,17 +402,6 @@ class OrderManagementController extends Controller
 						Cart::instance('Ordered_Bqt')
 				         ->add(['id' => $newBqt_Id, 'name' => $newBqtName, 'qty' => 1, 'price' => $BQT_Price,
 				         		'options' => ['count' => $BQT_Flower_Count]]);
-
-				       /* foreach(Cart::instance('Ordered_Bqt')->content() as $bqt){
-				    		echo '<div class = "row">';
-				    		echo '<h1>The Bouquet Detais<h1><hr><br>';
-				        	echo '<h2><b> $bqt_ID = </b>'.$bqt->id.'</h2>';
-				        	echo '<h4><b> $bqt_name = </b>'.$bqt->name.'</h4>';
-				        	echo '<h4><b> $bqt_qty = </b>'.$bqt->qty.'</h4>';
-				        	echo '<h4><b> $bqt_price = </b>'.$bqt->price.'</h4>';
-				        	echo '<h4><b> $bqt_count = </b>'.$bqt->options['count'].'</h4>';
-				        	echo '</div>';
-				    	}     */
 
 				        foreach(Cart::instance('OrderedBqt_Flowers')->content() as $row){
 						    //this foreach will transafer all of their content to another session
@@ -495,9 +419,7 @@ class OrderManagementController extends Controller
 			Cart::instance('OrderedBqt_Acessories')->destroy();
 	        Session::put('Save_Bouqet_To_myOrder', 'Successful');
           return redirect()-> route('Long_Sales_Order.index');
-        //return view('Orders.creationOfOrders')
-	     	//->with('FlowerList',$AvailableFlowers);//returns to the creation of orders
-     //}
+    }
 	}//end of function
 
 
@@ -509,21 +431,7 @@ class OrderManagementController extends Controller
             return redirect() -> route('adminsignin');
         }
         else{*/
-			/*$order_Details = DB::select('CALL salesorder_details(?)',array($order_ID));
-			$salesorder_Acessories = DB::select('CALL acessories_per_Specifici_Order(?);',array($order_ID));
-			$salesorder_Bouquet = DB::select('CALL salesorder_Bouquet(?)',array($order_ID));
-			$flowers_per_specific_BQTorder = DB::select('CALL flowers_per_specific_BQTorder(?)',array($order_ID));
-			$salesorder_Flowers = DB::select('CALL salesorder_Flowers(?)',array($order_ID));
-	          $Total_AmtBQT = DB::table('sales_order_bouquet')
-	          ->select(DB::raw('SUM(Unit_Price*QTY) as Amt'))
-	          ->where('Order_ID','=',$order_ID)
-	          ->get();
 
-	          $Total_AmtFLWR = DB::table('sales_order_flowers')
-	          ->select(DB::raw('SUM(Unit_Price*QTY) as Amt'))
-	          ->where('Sales_Order_ID','=',$order_ID)
-	          ->get();
-	*/
 	          $cities = DB::table('cities')
 	          ->select('*')
 	          ->get();
@@ -532,37 +440,24 @@ class OrderManagementController extends Controller
 	          ->select('*')
 	          ->get();
 
-			//return view('Orders.newConfirmation_of_Order');
-			/*return view('Orders.confirmation_of_Order')
-			->with('Total_AmtFLWR',$Total_AmtFLWR)
-			->with('Total_AmtFLWR1',$Total_AmtFLWR)
-			->with('Total_AmtBQT',$Total_AmtBQT)
-			->with('Total_AmtBQT1',$Total_AmtBQT)
-			->with('Flwr',$salesorder_Flowers)
-			->with('Bqt',$salesorder_Bouquet)
-			->with('Bqt_F',$flowers_per_specific_BQTorder)
-			->with('Order_det',$order_Details)
-			->with('Acrs',$salesorder_Acessories)
-			->with('city',$cities)
-	        ->with('province',$province);*/
 	        return view('Orders.confirmation_of_Order')
 			->with('city',$cities)
 	        ->with('province',$province);
-    	//}
+    	}
 	}//end of function
 
 	public function return_to_CreationOfOrder()
 	{//
-       /* if(auth::check() == false){
+       if(auth::check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
-			$AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
+        else{
+			       $AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
 
-			return view('Orders.creationOfOrders')
-	     	->with('FlowerList',$AvailableFlowers);
-	     //}
+			          return view('Orders.creationOfOrders')
+	     	         ->with('FlowerList',$AvailableFlowers);
+	     }
 	}//end of function
 
 }
